requirements.txt
apache-airflow==2.10.3
apache-airflow-providers-google==11.0.0
pydantic==2.10.2
pytest==8.3.4
requests==2.32.3
python-dotenv=1.0.1
